Include scroll arrow on landing, general styling changes + navbar
Chevron arrow added with bouncing animation and 'scroll' prompt to
direct users to scroll down. Animation added to chevron.
Background color changed to a gradient with height adjusted to allow
seamless design.
Navbar stylings added, with a script that controls when it will appear;
around the edge of the about and landing sections.